% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_data.R
\name{magmatize_data}
\alias{magmatize_data}
\title{Preparing MAGMA input data}
\usage{
magmatize_data(wd, age_classes = "all", loci_names = NULL, save_data = TRUE)
}
\arguments{
\item{wd}{Directory where you set up the \emph{data} folder.}

\item{age_classes}{Hard code class categories for group ages.}

\item{loci_names}{Optional. String containing loci names.}

\item{save_data}{Logical (with default = TRUE). Option to save the data in the \emph{data} folder.}
}
\value{
A list objects as the input data for \code{msgsi_mdl()}
}
\description{
Preparing MAGMA input data
}
\examples{
wd <- getwd() # path to data folder
magma_data <- magmatize_data(wd = wd, save_data = FALSE)

}
