% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_summ_bd_2.R
\name{magmatize_summ_bd2}
\alias{magmatize_summ_bd2}
\title{Summarize big daddy model output step 2}
\usage{
magmatize_summ_bd2(out1, summ_level, type)
}
\arguments{
\item{out1}{Output from step 1}

\item{summ_level}{Summarize at district or subdistrict level}

\item{type}{Identify "pop" or "age" to summarize only populations or age class.}
}
\value{
Model output in process and metadata as a list object.
}
\description{
Summarize big daddy model output step 2
}
\examples{
\dontrun{
# format data
wd <- getwd() # path to data folder
magma_data <- magmatize_data(wd = wd, save_data = FALSE)

# model run
magma_out <- magmatize_mdl(magma_data,
  nreps = 50, nburn = 25, thin = 1, nchains = 2)

# summary steps 1 and 2
tbr1 <- magmatize_summ_bd1(which_dist = 1,
  ma_out = magma_out,
  ma_dat = magma_data)

tbr2 <- magmatize_summ_bd2(tbr1, summ_level = "district", type = "pop")
}

}
