% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_summ_bd_3.R
\name{magmatize_summ_bd3}
\alias{magmatize_summ_bd3}
\title{Summarize big daddy model output ste 3}
\usage{
magmatize_summ_bd3(out2)
}
\arguments{
\item{out2}{Output from step 2}
}
\value{
Model output in process and metadata as a list object.
}
\description{
Summarize big daddy model output ste 3
}
\examples{
# format data
wd <- getwd() # path to data folder
magma_data <- magmatize_data(wd = wd, save_data = FALSE)

# model run
magma_out <- magmatize_mdl(magma_data,
  nreps = 50, nburn = 25, thin = 1, nchains = 2)

# summary steps 1 to 3
tbr1 <- magmatize_summ_bd1(which_dist = 1,
  ma_out = magma_out,
  ma_dat = magma_data)

tbr2 <- magmatize_summ_bd2(tbr1, summ_level = "district", type = "pop")

tbr3 <- magmatize_summ_bd3(tbr2)

}
