% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_summ_tbr_1.R
\name{magmatize_summ_tbr1}
\alias{magmatize_summ_tbr1}
\title{TBR output format wrapper step 1}
\usage{
magmatize_summ_tbr1(which_dist = NULL, ma_out, ma_dat)
}
\arguments{
\item{which_dist}{Function format raw magma output one district at a time.
Identify district as 1, 2, ... Default = NULL will summarize all districts.}

\item{ma_out}{MAGMA output.}

\item{ma_dat}{The same MAGMA input data for model run.}
}
\value{
Model output in multiway array and subset of metadata as a list object.
}
\description{
TBR output format wrapper step 1
}
\examples{
# format data
wd <- "D:/bobby_adfg/projects/magma/test_TBR" # path to data folder
magma_data <- magmatize_data(wd = wd, save_data = FALSE)

# model run
magma_out <- magmatize_mdl(magma_data,
  nreps = 50, nburn = 25, thin = 1, nchains = 3)

# summary step 1
tbr1 <- magmatize_summ_tbr1(which_dist = 3,
  ma_out = magma_out,
  ma_dat = magma_data)

}
