% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_summ_tbr_3.R
\name{magmatize_summ_tbr3}
\alias{magmatize_summ_tbr3}
\title{TBR output format wrapper step 3}
\usage{
magmatize_summ_tbr3(out2)
}
\arguments{
\item{out2}{Output from step 2}
}
\value{
Model output in process and metadata as a list object.
}
\description{
TBR output format wrapper step 3
}
\examples{
# format data
wd <- "D:/bobby_adfg/backup_013122/projects/magma/test_TBR" # path to data folder
magma_data <- magmatize_data(wd = wd, save_data = FALSE)

# model run
magma_out <- magmatize_mdl(magma_data,
  nreps = 50, nburn = 25, thin = 1, nchains = 3)

# summary steps 1 and 2
tbr1 <- magmatize_summ_tbr1(which_dist = 3,
  outraw = magma_out,
  ma_dat = magma_data,
  nreps = 50, nburn = 25, thin = 1, nchains = 3)

tbr2 <- magmatize_summ_tbr2(tbr1, summ_level = "district", type = "pop")

tbr3 <- magmatize_summ_tbr3(tbr2)

}
