% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_data.R
\name{prep_malia_data}
\alias{prep_malia_data}
\title{Prepare malia data}
\usage{
prep_malia_data(magma_data, path)
}
\arguments{
\item{magma_data}{Magma data object made using \code{magmatize_data()} function.}

\item{path}{A designated directory where you set up for malia data set.}
}
\value{
Items in magma data object saved as text files
}
\description{
Take a magma data object and save the items as text files in a designated directory.
}
\examples{
\dontrun{
prep_malia_data(magma_data, "D:/bobby_adfg/projects/magma/malia/data/bb2022")
}

}
