% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_plot.R
\name{tr_plot}
\alias{tr_plot}
\title{Plot MCMC trace}
\usage{
tr_plot(obj, nburn = 0, thin = 1, name_order = NULL)
}
\arguments{
\item{obj}{Trace from the model output.}

\item{nburn}{Number of burn-in you set up when you ran the model.
Default is 0 if you didn't save the burn-ins (keep_burn = FALSE).}

\item{thin}{Number of thinning you set up when you ran the model.
Default is 1 (no thinning).}

\item{name_order}{Arrange the reporting groups as you wish. Leave it empty
if you want to accept the default.}
}
\value{
Trace plot in ggplot
}
\description{
Plot MCMC trace
}
\examples{
# format data
wd <- "D:/bobby_adfg/backup_013122/projects/magma/test_TBR" # path to data folder
magma_data <- magmatize_data(wd = wd, save_data = FALSE)

# model run
magma_out <- magmatize_mdl(magma_data,
  nreps = 50, nburn = 25, thin = 1, nchains = 3, tbr = TRUE)

# summary
magma_summ <- magmatize_summ_tbr(which_dist = 2,
  outraw = magma_out,
  ma_dat = magma_data,
  nreps = 50, nburn = 25, thin = 1, nchains = 3,
  summ_level = "district")

# trace plot
tr_plot(obj = magma_summ$pop_prop[[1]])

}
