% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_indiv_assn.R
\name{magmatize_indiv}
\alias{magmatize_indiv}
\title{Individual group membership assignment summary}
\usage{
magmatize_indiv(ma_out, ma_dat, out_repunit = FALSE)
}
\arguments{
\item{ma_out}{MAGMA model output object name.}

\item{ma_dat}{MAGMA input data object name.}

\item{out_repunit}{Logical (default = \code{FALSE}). Option to output the likelihood in populations (\code{FALSE}) or combined in reporting groups (\code{TRUE}).
\itemize{
\item NOTE: combining populations into reporting group only works for single district or multi-district with the same reporting groups.
}}
}
\value{
A tibble contains posterior means of reporting group memberships for each individual of MAGMA metadata.
}
\description{
Individual group membership assignment summary
}
\examples{
\dontrun{
wd <- getwd() # path to data folder
magma_data <- magmatize_data(wd = wd, save_data = FALSE)
magma_out <- magmatize_mdl(magma_data, nreps = 50, nburn = 25, thin = 1, nchains = 2)
magma_assn <- magmatize_indiv(magma_out, magma_datt, out_repunit = TRUE)
}

}
