% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_indiv.R
\name{magmatize_indiv}
\alias{magmatize_indiv}
\title{Individual group membership assignment likelihood}
\usage{
magmatize_indiv(magma_dat, out_repunit = FALSE)
}
\arguments{
\item{magma_dat}{MAGMA data formatted using \code{magmatize_data()}.}

\item{out_repunit}{Logical (default = \code{FALSE}). Option to output the likelihood in populations (\code{FALSE}) or combined in reporting groups (\code{TRUE}).
\itemize{
\item NOTE: combining populations into reporting group only works for single district or multi-district with the same reporting groups.
}}
}
\value{
A tibble contains likelihood of reporting group memberships for each individual of MAGMA metadata.
}
\description{
Individual group membership assignment likelihood
}
\examples{
\dontrun{
wd <- getwd() # path to data folder
magma_data <- magmatize_data(wd = wd, save_data = FALSE)
scaled_like <- magmatize_indiv(magma_data)
}

}
