% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_mdl.R
\name{magmatize_mdl}
\alias{magmatize_mdl}
\title{Run MAGMA model}
\usage{
magmatize_mdl(
  dat_in,
  nreps,
  nburn,
  thin,
  nchains,
  nadapt = 50,
  keep_burn = FALSE,
  age_prior = "flat",
  cond_gsi = TRUE,
  file = NULL,
  seed = NULL
)
}
\arguments{
\item{dat_in}{Input data list.}

\item{nreps}{Amount of simulations to run.}

\item{nburn}{Number of burn-in simulations to discard.}

\item{thin}{At what interval to keep the simulations.}

\item{nchains}{Number of independent MCMC chains run in the simulation.}

\item{nadapt}{Amount of warm-up/adapt runs before the simulation (only for fully Bayesian mode).}

\item{keep_burn}{Logical (default = \code{FALSE}). To keep the burn-ins in the output or not.}

\item{age_prior}{Option to adjust prior weight on the age class proportions.
\itemize{
\item \code{flat} (default): conventional setup that puts a flat prior on the age proportions.
\item \code{zero_out}: prior weight will concentrate on the major age groups that are observed in metadata (i.e., "zero out" the unobserved age classes).
\item \code{weak_flat}: a less influential flat prior than the conventional one
}}

\item{cond_gsi}{Logical (default = \code{TRUE}). Option to use conditional GSI model. See vignette for details.
\itemize{
\item \code{TRUE}: run MAGMA with a hybrid algorithm of conditional GSI and fully Bayesian.
\item \code{FALSE}: run MAGMA with fully Bayesian algorithm.
}}

\item{file}{File path for saving the output in Rds file. Need to type out the full path, including file name and extension \code{.Rds}.
The default is \code{NULL} for not saving the output.}

\item{seed}{Option to initialize a pseudo-random number generator (set random seed)
so the output can be reproduced exactly.
Just pick a seed number and make note of it for future reference.
The default is \code{NULL}.}
}
\value{
A list object contains 1) the raw output of MAGMA as a list/multi-way array that need to be further summarized using summary functions, and 2) specifications for the model run (information needed for summary).
}
\description{
Run MAGMA model
}
\examples{
\dontrun{
# format data
wd <- getwd() # path to data folder
magma_data <- magmatize_data(wd = wd, save_data = FALSE)

# model run
magma_out <- magmatize_mdl(magma_data, nreps = 50, nburn = 25, thin = 1, nchains = 2)
}

}
