% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_summ.R
\name{magmatize_summ}
\alias{magmatize_summ}
\title{Output format wrapper}
\usage{
magmatize_summ(which_dist = NULL, ma_out, ma_dat, summ_level, type = NULL)
}
\arguments{
\item{which_dist}{Function format raw magma output one district at a time.
Identify district as 1, 2, ... Default = NULL will summarize all districts.}

\item{ma_out}{MAGMA output}

\item{ma_dat}{MAGMA input data}

\item{summ_level}{Summarize at district or subdistrict level}

\item{type}{Identify "pop" or "age" to summarize only populations or age class.
if you don't specify a "type", it will summarize both pop and age at the same time.}
}
\value{
Summary tables for reporting groups and/or age classes.
}
\description{
Output format wrapper
}
\examples{
# format data
wd <- "D:/bobby_adfg/projects/magma/test_TBR" # path to data folder
magma_data <- magmatize_data(wd = wd, save_data = FALSE)

# model run
magma_out <- magmatize_mdl(magma_data,
  nreps = 50, nburn = 25, thin = 1, nchains = 3)

# summary
magma_summ <- magmatize_summ(which_dist = 2,
  ma_out = magma_out,
  ma_dat = magma_data,
  summ_level = "district", type = "pop")

}
